[ admin ] Restore ORGANISING.md
This contained a bunch of useful information for first-time organisers
of SPLS, which is a semi-regular occurrence given that organising
occasionally gets delegated to some current PhD student.

I have made some adjustments, mostly just adjusting for post-covid-19
suggestions, but it is possible further adjustments are needed. Good
thing it is on git then!

The commit which removed it (d79db98) mentions that the file has been
moved to spls-organisation. However, I have been unable to find this
under the SPLI GitHub organisation, as well as the SPLI website. Hence
this restoration.